requirements.txt
Django==5.1.7
djangorestframework==3.16.0
pillow==11.1.0